requirements.txt
torchvision==0.3.0
numpy==1.16.4
torch==1.1.0
pandas==0.24.2
setuptools==39.0.1
python_geohash==0.8.5
catboost==0.15.1
matplotlib==3.1.0
seaborn==0.9.0
scikit_learn==0.21.2
